modulert ut navBarMaker, mindre kodekluss i Index
